Rename ESM test fixtures to .mjs
