+++ __tests__/fixtures-esm/tasks/wouldyoulike.mjs
@@ -1,4 +1,4 @@
-import { rand } from "../blah.js";
+import { rand } from "../blah.mjs";
 export default (_payload, helpers) => {
   helpers.logger.debug(rand());
   return "some sausages";
