+++ __tests__/fixtures-esm/tasks/wouldyoulike_default.mjs
