+++ jest.config.js
@@ -4,8 +4,8 @@ module.exports = {
     "^.+\\.tsx?$": "ts-jest",
   },
   testRegex: "(/__tests__/.*\\.(test|spec))\\.[tj]sx?$",
-  moduleFileExtensions: ["ts", "js", "json"],
-  extensionsToTreatAsEsm: [".ts"],
+  moduleFileExtensions: ["ts", "mjs", "js", "json"],
+  extensionsToTreatAsEsm: [".ts", ".mjs"],
   moduleNameMapper: {
     "^(\\.{1,2}/.*)\\.js$": "$1",
   },
